#69 Hard code IONOSPHERE_PROCESSES to 1 IssueID #2526: Hard code IONOSPHERE_PROCESSES to 1
This variable is only declared for the purpose of
maintaining a standard set up in each module and to possibly enable more
than one processor on Ionosphere in the future, should there be a
requirement for Ionosphere to analyse the metrics quicker.  Running
Ionosphere with more than one process is untested and currently it is
hard coded to be 1 (https://github.com/earthgecko/skyline/issues/69), fixes #69
and closes #69.

Modified:
skyline/ionosphere/ionosphere.py
skyline/settings.py